Eliminate User implementating TepidDb
It's not stored in the DB. This ought to be reason enough to not have it contain DB data.

Nothing which uses User uses the DB attributes, which is also a good reason to allow its removal.

It is also necessary for using DestinationTicket with Hibernate. This is necessary because User contains fields for the DB, but the class which wants to reference User also have those fields. This means that there are problems with all ways of using Hibernate for this task. That is: User is referenced in some other classes. In order to properly store that in a DB with Hibernate, User can either be marked as @Entity or @Embeddable. Marking it as @Entity means that it is treated as a separate class from FullUser. If those two were part of a hierarchy of classes, a polymorphic reference could be used (to extract just the User fields) . However, they aren't, meaning that there would now be parallel tables for User and FullUser. This creates all sorts of stupidity downstream, where a consumer of a User when handling a persisted FullUser must take on the burden of also ensuring that the FullUser is persisted as a User. The alternative is to mark User as @Embeddable. This causes the User to be stored along with the parent object. This is like the way it is currently used in the DB. However, since User inherits from TepidDb like the embedding class, it has all the DB metadata (such as ID) that the embedding class does, leading to a conflict. This conflict is irreconcilable, leading to a duplicate-fields error.

We could also collapse the *User classes, but a similar problem will be faced when using any of the other variants.