Use @Embedded instead of @EmbeddedId
I guess this is technically what I want? Although the embedded object contains the Id, it seems that @EmbeddedId treats the whole thing as the ID? and @Embedded will pull up the fields of the @Embeddable object into the table of the main @Entity, which is more like what we wanted
Maybe this fixes stuff too?